Fixed toggler and FREE card accessability
