+++ app/index.html
@@ -19,6 +19,7 @@
       href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@6.5.0/css/all.min.css"
     />
 </head>
+<script type="module" src="./js/main.js"></script>
 <body>
   <!-- <header>
     <div class="navbar navbar-dark bg-dark shadow-sm">
@@ -103,7 +104,8 @@
           </form>
         </div>
         <button id="theme_toggle" class="btn">
-          <i class="fa-solid fa-moon"></i>
+          <i id="toggle_dark" class="fa-solid fa-moon"></i>
+          <i id="toggle_light" class="fa-solid fa-sun d-none"></i>
         </button>
       </div>
