+++ app/js/main.js
@@ -1,0 +1,5 @@
+import ThemeToggle from './ui/ThemeToggle.js';
+
+window.addEventListener('DOMContentLoaded', () => {
+  new ThemeToggle();
+});
