+++ app/js/ui/ThemeToggle.js
@@ -1,0 +1,29 @@
+export default class ThemeToggle {
+  constructor(buttonId = 'theme_toggle') {
+    this.button = document.getElementById(buttonId);
+    this.root = document.documentElement; // <html>
+
+    if (!this.button) return;
+
+    this.button.addEventListener('click', () => this.toggleTheme());
+  }
+
+  toggleTheme() {
+    const current = this.root.getAttribute('data-bs-theme');
+    const newTheme = current === 'dark' ? 'light' : 'dark';
+    this.root.setAttribute('data-bs-theme', newTheme);
+    this.updateIcon(newTheme);
+  }
+
+    updateIcon(theme) {
+      const darkIcon = document.getElementById('toggle_dark');
+      const lightIcon = document.getElementById('toggle_light');
+    if (theme === 'dark') {
+      darkIcon.classList.remove('d-none');
+      lightIcon.classList.add('d-none');
+    } else {
+      lightIcon.classList.remove('d-none');
+      darkIcon.classList.add('d-none');
+    }
+  }
+}
